feat: add integration test for Spring Redis application
A new file `SpringRedisIT.java` was added to the `src/test/java/com/mvr/example` directory. This file contains an integration test for the Spring Redis application.

The integration test is annotated with `@SpringBootTest` and `@AutoConfigureMockMvc`. It sets the `webEnvironment` to `MOCK` and specifies the `SpringRedisApplication` class as the configuration class.

The test class also includes an `Autowired` field `mvc` of type `MockMvc`, which will be used for performing HTTP requests and asserting the responses in the test cases.